backend/requirements.txt
fastapi
uvicorn
python-dotenv
web3
qrcode
google-generativeai
pydantic
